% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.network.R
\name{make.network}
\alias{make.network}
\title{Take an edgelist and create a network of simfrogs with relevant attributes.}
\usage{
make.network(df, adj.matrix, temp, noiselevel)
}
\arguments{
\item{df}{dataframe generated by uniform.frog.sex.ratio.setup}

\item{adj.matrix}{adjacency matrix generated by matrixify}

\item{temp}{numeric value of temperature in degrees celsius}

\item{noiselevel}{numeric value of noise level used in matrixify}
}
\value{
network of frogs
}
\description{
Take an edgelist and create a network of simfrogs with relevant attributes.
}
\seealso{
\code{\link{matrixify}} which this function depends on
}
