% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/male.ranker.R
\name{male.ranker}
\alias{male.ranker}
\title{Ranks males by their attractiveness and removes connections to the lowest-ranked males.}
\usage{
male.ranker(g)
}
\arguments{
\item{g}{network object generated by generate.uniform.frogs}
}
\value{
network object
}
\description{
Ranks males by their attractiveness and removes connections to the lowest-ranked males.
}
\seealso{
\code{\link{generate.uniform.frogs}} upon which this function is dependent
}
