% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixify.R
\name{matrixify}
\alias{matrixify}
\title{Create an adjacency matrix of simulated frogs and remove any edges where the frogs can't hear each other.}
\usage{
matrixify(df, noiselevel)
}
\arguments{
\item{df}{dataframe generated by uniform.frog.sex.ratio.setup}

\item{noiselevel}{numeric value in dB of ambient noise level at simulated pond}
}
\value{
edgelist of frogs
}
\description{
Create an adjacency matrix of simulated frogs and remove any edges where the frogs can't hear each other.
}
\seealso{
\code{\link{uniform.frog.sex.ratio.setup}} which this function depends on
}
