% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.corrected.density.R
\name{sex.corrected.density}
\alias{sex.corrected.density}
\title{Sex-corrected density measure for only finding the male-female connections.}
\usage{
sex.corrected.density(g)
}
\arguments{
\item{g}{network object}
}
\value{
numeric sex-corrected density
}
\description{
Sex-corrected density measure for only finding the male-female connections.
}
